+++ tests/testthat/test-info.R
@@ -78,3 +78,17 @@ test_that("mc_info_join", {
     expect_warning(table2 <- mc_info_join(merged_data))
     expect_equal(table2$count_errors, 1)
 })
+
+test_that("mc_info_states", {
+    data <- mc_read_data("../data/TOMST/files_table.csv", silent=TRUE)
+    states <- mc_info_states(data)
+    expect_equal(colnames(states), c("locality_id", "logger_index", "logger_type", "sensor_name",
+                                     "tag", "start", "end", "value"))
+    expect_equal(nrow(states), 9)
+    agg_data <- mc_agg(data, "max", period="hour")
+    states <- mc_info_states(agg_data)
+    expect_equal(colnames(states), c("locality_id", "logger_index", "logger_type", "sensor_name",
+                                     "tag", "start", "end", "value"))
+    expect_equal(nrow(states), 9)
+    expect_true(all(is.na(states$logger_index)))
+})